[NEW] Scripts d'automatisation du build [ENHANCE] Localisation de certains messages d'erreurs non affichés jusqu'à présent [FIX #25449] Problèmes dans l'installeur web

+++ build.sh
@@ -1,5 +1,19 @@
+#!/bin/bash
+
+if [ ! -d themes/webfactory ]; then
+	echo "Missing themes/webfactory directory. Please run get-theme.sh"
+	exit 1
+fi
+
+if [ ! -d repository ]; then
+	echo "Missing repository directory. Please run gen-repository.sh"
+	exit 1
+fi
+
 rm -rf build
 mkdir build
+
+cp pack/change.properties change.properties
 
 cp pack/cmscore.change.xml change.xml
 tar --exclude .git -czf build/cmscore-downloader-3.0.4.tgz change.properties change.xml profile config install themes index.php
